Demo: organise view components under app/views
Move BarView into app/views/bar_view.rb and add a reusable ButtonView
(app/views/button_view.rb), each required by the scene. ReactiveScene's Back
and Add buttons now share ButtonView(id:, label:, action:), showing components
living in their own files and composed into a scene's view.
