+++ demo/mygame/app/scenes/reactive_scene.rb
@@ -1,28 +1,6 @@
-# A view component for one progress bar row. Invoked as BarView(item:, ...) from
-# the scene's view. #call emits the row's nodes; clicking the row calls back into
-# the scene via the on_remove prop.
-class BarView < Conjuration::UI::View
-  BAR_MAX = 340
-
-  def initialize(item:, width:, on_remove:)
-    @item = item
-    @width = width
-    @on_remove = on_remove
-  end
-
-  def call
-    id = @item[:id]
-    progress = @item[:progress]
-    remove = @on_remove
-
-    node({ w: @width, h: 26, direction: :row, gap: 12, align: :center, action: -> { remove.call(id) } }, id: "row_#{id}") do
-      node({ w: 48, h: 26, justify: :center, align: :center }, id: "pct_#{id}") do
-        node({ text: "#{progress.to_i}%", r: 210, g: 210, b: 220 })
-      end
-      node({ w: BAR_MAX * progress / 100, h: 18, path: :pixel, r: 120, g: 200, b: 120 }, id: "fill_#{id}")
-    end
-  end
-end
+# View components live under app/views and are required per scene.
+require_relative "../views/button_view"
+require_relative "../views/bar_view"
 
 # A reactive scene demonstrating the view reconciler. The HUD is declared once
 # as `view`, a pure function of state. update() only animates bar[:progress];
@@ -65,12 +43,8 @@ class ReactiveScene < Conjuration::Scene
 
   def view
     node({ x: 20, y: 20.from_top, anchor_y: 1, direction: :row, gap: 12 }, group: :controls) do
-      node({ w: 100, h: 44, path: "sprites/button.png", action: -> { change_scene(to: MenuScene.new(:main)) } }, id: :back, justify: :center, align: :center) do
-        node({ text: "Back", r: 255, g: 255, b: 255 })
-      end
-      node({ w: 140, h: 44, path: "sprites/button.png", action: -> { spawn } }, id: :add, justify: :center, align: :center) do
-        node({ text: "Add bar", r: 255, g: 255, b: 255 })
-      end
+      ButtonView(id: :back, label: "Back", action: -> { change_scene(to: MenuScene.new(:main)) })
+      ButtonView(id: :add, label: "Add bar", width: 140, action: -> { spawn })
     end
 
     node({ x: grid.w / 2, y: grid.h / 2, w: PANEL_WIDTH, h: 360, anchor_x: 0.5, anchor_y: 0.5, path: :pixel, r: 24, g: 26, b: 34 }, id: :panel, gap: 10, padding: 20, justify: :center) do
